+++ DefaultOffenseProcessor/src/main/java/me/superbiebel/defaultoffenseprocessor/DefaultOffenseProcessor.java
@@ -1,6 +1,6 @@
 package me.superbiebel.defaultoffenseprocessor;
 
-import me.superbiebel.punishmentmanager.offenseprocessing.OffenseProcessor;
+import me.superbiebel.punishmentmanager.offenseprocessing.abstraction.OffenseProcessor;
 
 public class DefaultOffenseProcessor implements OffenseProcessor {
 }
